kv: Move in some metrics
