+++ src/server/metrics.rs
@@ -6,6 +6,8 @@ use prometheus_static_metric::*;
 use crate::storage::ErrorHeaderKind;
 use prometheus::exponential_buckets;
 
+pub use crate::storage::kv::metrics::{GcKeysCF, GcKeysDetail, GcKeysCounterVec, GcKeysCounterVecInner};
+
 make_auto_flush_static_metric! {
     pub label_enum GrpcTypeKind {
         invalid,
@@ -77,26 +79,6 @@ make_auto_flush_static_metric! {
         tombstone,
     }
 
-    pub label_enum GcKeysCF {
-        default,
-        lock,
-        write,
-    }
-
-    pub label_enum GcKeysDetail {
-        processed_keys,
-        get,
-        next,
-        prev,
-        seek,
-        seek_for_prev,
-        over_seek_bound,
-        next_tombstone,
-        prev_tombstone,
-        seek_tombstone,
-        seek_for_prev_tombstone,
-    }
-
     pub struct GcCommandCounterVec: LocalIntCounter {
         "type" => GcCommandKind,
     }
@@ -119,11 +101,6 @@ make_auto_flush_static_metric! {
 
     pub struct GrpcMsgFailCounterVec: LocalIntCounter {
         "type" => GrpcTypeKind,
-    }
-
-    pub struct GcKeysCounterVec: LocalIntCounter {
-        "cf" => GcKeysCF,
-        "tag" => GcKeysDetail,
     }
 
     pub struct GrpcMsgHistogramVec: LocalHistogram {
